pyproject.toml
[project]
name = "windowsserverhoneypot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncssh>=2.21.0",
    "dnslib>=0.9.26",
    "impacket>=0.12.0",
    "loguru>=0.7.3",
    "pyyaml>=6.0.2",
]
